Accidentally put this on the Splay_tree branch when I meant to only put it on the Master_For_Everything Added more to AVL | Splay is still starting
Most of it I am not sure if it is correct or not and most of it needs to be heavily tested. I am not even sure if one method in there works the way it is supposed to or not.
Element Done
Node Done
BST Done
AVL in Progress
SplayTree in Progress