Still Stuck in Splay_Tree and dont know how to change branches.
Added more comments to AVL and Splay

Added more to AVL confused as heck | Splay is still confused as heck
AVL needs heavy testing and checking
Splay needs writing, testing, and then checking

Most of it I am not sure if it is correct or not and most of it needs to be heavily tested. I am not even sure if one method in there works the way it is supposed to or not.

Element Done
Node Done
BST Done
AVL in Progress
SplayTree in Progress